move multi agent mlp
